# Optimize stub_with